+++ tests/testthat/test-chop.R
@@ -162,9 +162,12 @@ test_that("chop_n", {
 
 test_that("chop_mean_sd", {
   x <- -1:1 # mean 0, sd 1
-  expect_silent(chop_mean_sd(x))
-  expect_silent(chop_mean_sd(x, sd = 2))
-  expect_silent(chop_mean_sd(x, sd = 1.96))
+  expect_silent(res <- chop_mean_sd(x))
+  expect_equivalent(as.vector(table(res)), c(1, 1, 1))
+  expect_silent(res2 <- chop_mean_sd(x, sds = 1:2))
+  expect_silent(chop_mean_sd(x, sds = c(1, 1.96)))
+  lifecycle::expect_deprecated(res3 <- chop_mean_sd(x, sd = 2))
+  expect_equivalent(res2, res3)
 })